Adds another way (edge detection) to calculate the orientation of the goal points.
If the goal point lies outside of the defined sensor range an edge detection
is performed. If a valid edge has been found (point lies on the edge)
there are two possible orientations: known -> unknown and unknown -> known.
The correct orientation is finally defined by counting unknown pixels in
both directions.

If problems occur finding the correct edge, the old way to
calculate the orientation (robot - goal vector) is used.